사용자의 답을 볼과 스트라이크 낫싱 체크하는 함수를 game domain에 이동
